backend/flask/requirements.txt
celery==4.0.2
Flask==0.12
Flask-Bcrypt==0.7.1
Flask-Migrate==2.0.3
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
ipython==5.3.0
psycopg2==2.7.1
PyJWT==1.4.2
